Add a new flag "lift" to the function disasm_inst_adv()
With the addition of a "lift" flag, the function will now optionally
"lift" pseudo-instructions from the raw instructions. Lifting is where
a raw instruction is converted to a more human-friendly one, for example
the raw instruction "jalr x0,0(x1)" would be lifted to be "ret".

Prior to this commit pseudo-instructions were always lifted. This
behaviour can still be achieved by passing any non-zero value as "lift".
Lifting can be suppressed by passing in the value zero as "lift"

Also, only try and write to the buffer "buf", if it is not NULL!

Signed-off-by: Sean McGoogan <Sean.McGoogan@ultrasoc.com>